+++ src/main/java/com/select/choice/domain/post/domain/data/response/PostResponse.java
@@ -1,6 +1,5 @@
-package com.select.choice.domain.post.domain.data.reponse;
+package com.select.choice.domain.post.domain.data.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
