+++ src/main/java/com/select/choice/domain/post/domain/util/PostConverter.java
@@ -3,7 +3,7 @@ package com.select.choice.domain.post.domain.util;
 import com.select.choice.domain.post.domain.data.dto.CreatePostDto;
 import com.select.choice.domain.post.domain.data.dto.PostDto;
 import com.select.choice.domain.post.domain.data.entity.Post;
-import com.select.choice.domain.post.domain.data.reponse.PostResponse;
+import com.select.choice.domain.post.domain.data.response.PostResponse;
 import com.select.choice.domain.post.domain.request.CreatePostRequestDto;
 
 import java.util.List;
